move inline parser codes to paragraph parser
